Move the cogs into a subdir
